Create Reducer Directory
Create a subfolder in the new reducer directory for restaurants.
Refactor store code to handle the new file structure
