+++ src/store/reducers/index.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import restaurantReducer from './restaurant';
+
+export default combineReducers({
+  restaurants: restaurantReducer
+});
